Gazebo plugin that listens to a Gazebo topic
